+++ src/test/TestAlignmentTraceback.cpp
@@ -169,8 +169,8 @@ int main (int argc, const char * argv[])
     dbSeq->mapSequence(1,"lala2",tim2.c_str());
     SmithWaterman aligner(15000, subMat.alphabetSize, false);
     int8_t * tinySubMat = new int8_t[subMat.alphabetSize*subMat.alphabetSize];
-    for (size_t i = 0; i < subMat.alphabetSize; i++) {
-        for (size_t j = 0; j < subMat.alphabetSize; j++) {
+    for (int i = 0; i < subMat.alphabetSize; i++) {
+        for (int j = 0; j < subMat.alphabetSize; j++) {
             std::cout << ( i*subMat.alphabetSize + j) << " " << subMat.subMatrix[i][j] << " ";
 
             tinySubMat[i*subMat.alphabetSize + j] = (int8_t)subMat.subMatrix[i][j];
@@ -180,7 +180,7 @@ int main (int argc, const char * argv[])
     short ** profile = new short*[subMat.alphabetSize];
     short * profile_data = new short[10000*subMat.alphabetSize];
 
-    for(size_t i = 0; i< subMat.alphabetSize; i++) {
+    for(int i = 0; i< subMat.alphabetSize; i++) {
         profile[i] = &profile_data[i*s->L];
         for (size_t j = 0; j < s->L; j++) {
             profile[i][j] = tinySubMat[i*subMat.alphabetSize + s->int_sequence[j]];
